Track Coc extensions and move .gitignore
Installed coc extensions are saved in a json file. To make the dotfiles
more portable, this file needs to be tracked. Once coc is installed,
`:CocInstall` will pull the extensions saved in package.json.

This commit also includes deletion of some whitespaces in init.vim and
moving coc keymapping to keymap section.